prov/efa: Only initialize devices that match FI_EFA_IFACE filter
FI_EFA_IFACE is used to restrict the EFA devices that are visible to the
application. There is no need to initialize devices that the application
does not want to use.

However, we still want to keep track of all GIDs on the instance even if
they are not selected in FI_EFA_IFACE because we use that list to
determine if a peer is local and create SHM provider resources for such
local peers.

So this commit splits efa_device_construct into two parts: the first
part initializes efa_device fields up to and including the GID. The
second part initializes the protection domain and prov_info structs. The
second part is only executed for EFA devices that match the FI_EFA_IFACE
filter.

And it introduces g_efa_ibv_gid_list and g_efa_ibv_gid_cnt to keep track
of all EFA device GIDs on the instance irrespective of FI_EFA_IFACE

Signed-off-by: Sai Sunku <sunkusa@amazon.com>