prov/efa: Move address handle to domain level.
Today, the efa address handle (AH) resources are
created and tracked in the AV level. This is wrong because for multiple AVs created in the same domain, they share the same PD and the created AHs will not be independent. When one AV gets closed and destroyed its AHs, WQEs posted via other AHs that share the same PD and GID will get aborted.

This patch fixes this issue by doing the following:

1. Move ah_map (hash map, key is GID, value is AH) from efa_av to efa_domain which now has a 1:1 map to PD. The map is protected by the domain lock.
2. When fi_av_insert decide to create a new av entry, look up the AH in the device ah_map (in the locking), if it's not found, create AH and initialize its refcnt as 1 and insert to hashmap. If it's found, return the AH and increment its refcnt.
3. In fi_av_remove, do the look up again in ah_map, decrement the refcnt of the returned AH. If the refcnt is 0, destroy the AH and delete it from the hashmap
4. In efa_domain close, iterate the ah map again and destroy all uncleaned AHs (with warning).

Signed-off-by: Shi Jin <sjina@amazon.com>