prov/sockets: fix rx_buffered arrived befor fi_av_insert
When an unexpected message arrive before fi_av_insert, the av_index
won't be available and was set to FI_ADDR_NOTAVAIL. It eventually may
get mismatched to a wrong recv entry. Store the raw ip address for this
case and try to recover the av_index at match time.

Signed-off-by: Hui Zhou <hzhou321@anl.gov>