include/ofi_atomic_queue: add entry init function
Add function definition to be able to initialize fields in
in the queue. This function is eager so the entry is already
initialized when it gets assigned to the caller and gets
pre-emptively re-initialized on release back into the queue.
This can help with caching if initialization is more effective
done by the owner instead of peers

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>