Fabtests: Add New Benchmark rma_tx_completion
This commit adds a new benchmark rma_tx_completion, which measures the
time for a TX completion to occur on the writer. The test does not do
any data validation.  The test sets xfers_per_transfer to 1 to not have
show_perf divide the latency result by 2 (the test is uni-directional
already).

Signed-off-by: Seth Zegelstein <szegel@amazon.com>