prov/util: Allow util provider to lock the info list
With the FI_RESCAN flag, a provider can now dynamically rescan
the list of interfaces. As a result, the util provider needs a
mechanism to protect the info list during iteration.

This patch allows any provider to register a pointer to a mutex
in util_prov, which is used to protect the info list.

Signed-off-by: Sylvain Didelot <sdidelot@ddn.com>