[v2.1.x] prov/efa: Remove illegal ope state
EFA_RDM_OPE_FREE is not a legal state.
It is currently set before the ope is freed,
but we should not access the ope after
its freed. This patch removes this
enum and its references. It keeps the enum
of EFA_RDM_TXE_REQ to be 1 because otherwise
the fresh allocated OPEs will have a state
as EFA_RDM_TXE_REQ (0).

Signed-off-by: Shi Jin <sjina@amazon.com>
(cherry picked from commit e78d06dec9493cf5573606b8bc8b2ec2b75ed195)