prov/shm: acquire ep lock when freeing entries
When closing the shm ep, we have to free entries with the owner srx if there
are any outstanding unexpected messages. When shm is the owner, that means we have
to lock the srx (uses the ep lock) to avoid lock debug errors. When used as a peer
it is expected that the owner lock the srx a peer EP is bound to when closing the
endpoint

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>