prov/efa: Bugfix, Return FI_ADDR_NOTAVAIL when FI_SOURCE not set
The libfabric man pages say:

If fi_cq_readfrom is called on an endpoint for which source addressing data is not available,
the source address will be set to FI_ADDR_NOTAVAIL.

The EFA provider funnels both cq read and readfrom calls into the
util_cq, and passes the util_cq a progress function. The util_cq calls
the progress function for both read/readfrom, and it is the progress
functions job to poll the endpoint[s] associated with the CQ. Since some
endpoints can be configured with FI_SOURCE, and others may not, the
util_cq has no ability to handle this check for us, and it must be done
inside the endpoint progress code.

This patch switches efa to always call ofi_cq_write_src to make sure
that the addr will always be set correctly. Without this change, the
addr will not be set and will default to its previous value.

Signed-off-by: Seth Zegelstein <szegel@amazon.com>