prov/efa: Grab srx lock when processing write with imm CQE in efa-direct
The local peer may never know about a pending write with immediate
operation and it may remove the peer from the AV while the peer is
performing a write with immediate operation.

Acquiring the srx lock will serialize the AV reverse lookup with AV
control plane operations and prevent the seg fault.

Signed-off-by: Sai Sunku <sunkusa@amazon.com>