prov/efa: Do not add explicit MRs from application to MR cache
This commit ensures that the MR cache is only used for internal MRs
registered by the EFA provider. The domain's fi_ops_mr is always set to
never use the MR cache. The fi_ops_mr with the cache is removed.

This commit adds a unit test that verifies that the MR cache is not
used for explicit MRs when FI_MR_LOCAL is unset. It also adds fabtests
that run fi_rdm_pingpong without FI_MR_LOCAL both with and without the
MR cache enabled.

Signed-off-by: Sai Sunku <sunkusa@amazon.com>