src/Makefile: Add version and commit startup log
This commit modifies the build system, as well as src/fabric.c and
src/fi_version_gen.h.in to create a info log which outputs the version
of libfabric that is running as well as the commit used to build.  The
log message looks like:

libfabric:906569:1749843568::core:core:fi_ini():1014<info> Initialized \
libfabric 2.2.0rc1 - built from commit 7ef321d3

This log message will be helpful when trying to prove which
version/commit of libfabric is being run. This is helpful when there are
multiple versions installed on the system, and application startup is
non-trivial (docker/enroot/slurm/etc...).

The src/fi_version_gen.h file gets generated on every make command
(including make dist) if there is a .git folder in the source directory.
The file is compiled into the library so it works for both .a/.so libs.
This has implications for the development workflow and distribution
users. When developers create a new commit, and rebuild it, the log
message will be updated. When developers bump the libfabric version in
configure.ac and configure/make, the log message will be updated. When a
developer wishes to release, and they call ./autogen.sh && ./configure
&& make dist, they will get a src distribution tarball.  They can then
go into that tarball and call configure/make and they will always get
the version/commit that was used to make the tarball.

When the user calls make clean in a git repo src dir, the generated
src/fi_version_gen.h gets deleted.  When the user calls make clean in a
dist src dir (no .git), the src/fi_version_gen.h file does not get
deleted.  This is required because the project will not build without
src/fi_version_gen.h, and in a dist src dir, the file does not get
generated.

Signed-off-by: Seth Zegelstein <szegel@amazon.com>