prov/efa: Add missing dlist removal in tx pke release
For tx pkt, pkt_entry->entry is linked to either
peer->outstanding_tx_pkts or txe->queued_pkts.
It must be removed from the list as part of the
pkt entry release, otherwise it will cause double
free bug when txe_handle_err and txe_release are
both called for the same txe.

Signed-off-by: Shi Jin <sjina@amazon.com>