prov/efa: Move rxe_map to peer struct
With an implicit AV, fi_addr cannot be used to identify unique peers.
All unknown peers will have fi_addr set to -1. And the index in the
implicit AV can be re-used, so it's not guaranteed to be unique.

This commit makes the rxe_map part of the peer struct and only uses the
msg_id as the hash key

Signed-off-by: Sai Sunku <sunkusa@amazon.com>