prov/shm: add shm rename and retry
The shm provider does its best to randomize an EP name and make it
unique by defaulting to "pid:domain_id:ep_id"
If there already exists a region with that name, the provider wil try
to map to the existing region and see if the pid is alive and remap if
the pid is dead.
This method doesn't work well with docker containers as the pids live
within the container. This means you can have a docker pid 100 that coincides
with a host system pid 100 running a root service OR you could have two docker
containers running simultaneously that are trying to use the same pid (and
therefore shm region name). This can result in a failure to create the shm
region or failure to detect whether the pid is actually alive.
To fix this, we can fallback to a retry of the shm create and append a random
seed number to the shm region name to make sure it is unique in case a shm
region already exists with that name.
There is a scenario where the app overrides the default shm name with a specific
name. If that's the case, we shouldn't override the user-selected name and,
instead, fail if it's not available.

cherry-picked from commit 0fb02def33cf1ad6551a7b775f285bfb6c5c2dcb

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>