prov/shm: fix srx entry cleanup
When cleaning up the srx (util_srx_close), the srx iterates through
the list of unexpected messages and calls discard_msg/tag on each
entry which means shm shouldn't have to iterate through the list of
unexpected messages and free each one. The unexp_cmd_list was only
there for cleanup so it can get removed.
The special case is unexpected SAR entries which also live in the sar
list so the pend entry can get cleaned up. In this case, we can't
free the srx entry but need to free the pend entry. This case is handled
in the sar list cleanup so only the expected messages are freed. For
the unexpected case, this will get triggered on the srx close

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>