prov/efa: Drain cq after qp destroy
After QP is destroyed, there can be completions in the CQ
belong to this QP tha are not polled yet. These stale completions
should be drained (as EINVAL from ibv_start_poll/ibv_next_poll)
otherwise it can be polled as active completions after a new QP
with the same QPN is created.

This patch implements this cq drain after the QP destruction
as part of the EP close in libfabric.

Signed-off-by: Shi Jin <sjina@amazon.com>