prov/efa: Implement LRU eviction for implicit AV entries
This commit introduces an LRU list for the implicit AV entries. The most
recently inserted/accesssed AV entry is at the tail of the list and the
least recently inserted/accessed AV entry is at the head of the list.
When the size of the list reaches the maximum size, the least recently
used entry at the head of the list is evicted.

Signed-off-by: Sai Sunku <sunkusa@amazon.com>