Revert "prov/efa: Remove fallback logic for determining peer address"
This reverts commit a13594de391f7f9daa7ac7b48bcd057b235b274c.

The EFA device can still return a GID of -1 for a peer that has already
been inserted into the AV. So we still need the fallback logic in some
cases.

This commit is required for the implicit AV change to work.

There were no test failures because inserting twice into the explicit AV
is idempotent. But with an implicit AV, attemption to insert an address
already in the explicit AV into the implicit AV will cause failures.

Signed-off-by: Sai Sunku <sunkusa@amazon.com>