include/fi_peer.h, prov/util,lnx: remove fi_peer_rx_entry dlist fields
fi_peer_rx_entry has two internal fi_peer_rx_entry fields so the struct could
be castable to a dlist_entry or slist_entry. This suggests a dependency between
the external headers and the internal headers which if difficult to ensure.
In addition, there is no real reason for the owner and peer to share an entry
for tracking the peer entry. The owner and peer each have their own structs and
mechanisms for tracking the entry when the resource is on their side so the
extra fields can be removed completely.

Since there is a requirement for all peers/owners to use the same version of
libfabric, there is no issue with backwards compatibility

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>