Revert "prov/efa: Only do dmabuf reg when FI_MR_DMABUF is set"
This patch reverts an earlier change that aims to do dmabuf reg only when FI_MR_DMABUF flag is set.
We still want to use dmabuf for neuron by default even when this flag is not set.

It also adds the clean up of the dmabuf fd after the ibv_reg_dmabuf_fd to fix the fd leak

This reverts commit 043e968534491319d1b9afc22e482783d5779677.

Signed-off-by: Shi Jin <sjina@amazon.com>