prov/cxi: correct RE arming behavior
The first time a particular reduction ID is used is not the only
time we may need to arm the reduction engine. We also need to re-
arm if the 'next' reduction is initiated after the previous arming
period has expired. Replace the "first send" logic with "need to
arm" logic.

Signed-off-by: John Heemstra <john.heemstra@hpe.com>