examples: add example directory with first example
Add a new examples/ directory which is intended to have standalone,
single-file examples that show how to use libfabric and its features.
A single Makefile is provided for simplicity in building.
Each test will be completely contained in a single file with no
cross depdencies and will be thoroughly commented to explain each
step in the program.
Each test executable will have the prefix fi_example_

This commit comes with the first example - fi_example_rdm which outlines
a simple server-client and send-recv

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>