prov/opx: Only open IPC cache if HMEM initialized and IPC enabled
An OPX build with ROCR resulted in a segmentation fault for a CPU only memory workload because of an error opening the IPC cache when ROCR IPC is enabled, but ROCR HMEM was not initialized. The log shows:
	opx:domain:opx_hmem_open_domain():191<warn> Error opening IPC Cache ret=-38 (Unknown error -38).

The OPX open domain path has been updated to only open the IPC cache if HMEM is initialized and IPC is enabled.

Signed-off-by: Lindsay Reiser <lindsay.reiser@cornelisnetworks.com>