include/ofi_indexer: introduce new ofi_array_at_max function
ofi_array_at is written so that the array will grow to the size
of the given index and return an always valid index so it can be
called on initialization of an entry.
However, if the caller is looking to return an already valid and
initialized entry, this new call can be used. This ensures that
the returned entry is a previously allocated and initialized entry
and will never grow the array.

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>