prov/util: allow user alter for domain control_progress
domain control_progress attribute was added earlier to
provide a lockless statemachine in libfabric with FI_THREAD_DOMAIN
threading and FI_PROGRESS_CONTROL_UNIFIED control_progress mode.
However there was no way for a user to provide the required
control_progress attribute. This patch is adding the support for
allowing user to override the control_progress along with the other
domain attributes.

Signed-off-by: Sunita Nadampalli <nadampal@amazon.com>