prov/util: Remove peer rx entry if it's in the unspec_unexp queue
The EFA provider AV keeps track of peers that any endpoint bound to the
AV has received messages from but were not explicitly inserted into the
AV by the application. The provider limits the number of such peers to
prevent runaway memory usage.

When the EFA provider has to remove such an implicitly inserted peer,
it also needs to remove any unmatched entries in the SRX queues
associated with the peer. The EFA provider calls free_entry to remove
the rx entry. But free_entry does not remove the rx entry from the
unspec_unexp_[msg,tag]_queue dlist which corrupts the dlist.

This commit adds logic to remove any entry that is about to be freed
from the unspec_unexp_[msg,tag]_queue.

Signed-off-by: Sai Sunku <sunkusa@amazon.com>