prov/efa: Ignore stale cqes from destroyed qps
Currently, efa provider always write cq errors when ibv_start_poll/ibv_next_poll
returns EINVAL, which means the cqes are from destroyed QPs (not found in
the qp table). This behavior is not complying with libfabric man page that
said the outstanding operations posted to the endpoint when fi_close is called
will be silently dropped without generating completions.

Signed-off-by: Shi Jin <sjina@amazon.com>