[2.3.x] prov/cxi: reset timer on rx of ARM packet
We were not (re)setting the timer for the "need to re-arm"
logic on the leaf side. This resulted in a performance
blip when the timer eventually expired. We now reset this
timer as soon as the leaf sees an ARM packet.

Signed-off-by: John Heemstra <john.heemstra@hpe.com>
(cherry picked from commit cfaac7afe6adfa6bb6e8f5c04d479ff6f039355a)
Signed-off-by: James Swaro <james.swaro@hpe.com>