prov/efa: Fix the memleak for efa_rdm_peer
Currently, the peer cleanup only happen in av remove or
av close. When application closes ep before closing av, such cleanup
won't happen because the ep is not in the av->ep_list any more.
This is actually wrong because peer is now a ep level resource
which should be cleaned as part of the ep close. This patch
fixes this issue.

Signed-off-by: Shi Jin <sjina@amazon.com>