prov/efa: allocate peer map entry pool during the rdm ep create
the commit "4016fd5" has moved the peer map entry pool from av to
endpoint but deferred the pool memory allocation till the remote
requests for a peer. So, the acutal memory pool allocation was
happening on the first request and hence adding latency overhead
to the intial requests.

This commit is fixing the letency issue by eagerly allocating the pool
during the endpoint create.

Signed-off-by: Sunita Nadampalli <nadampal@amazon.com>