[v2.3.x] prov/opx: Use getpid() instead of gettid() for POSIX compliancy.
The 'gettid()' function is not POSIX compliant, and older versions
of glibc may not have that function available in unistd.h, causing
a compile failure. This changes the call to 'getpid()', which is
POSIX compliant and should always be present in glibc.

Signed-off-by: Ben Lynam <Ben.Lynam@cornelisnetworks.com>
(cherry picked from commit 7660f36eccace6380c6ae310dadd76c6d0c39b8a)