[v2.3.x] prov/verbs: Fix compiler warning on the bound of 'strncpy' call
Gcc 15 complains that the length parameter for strncpy() is derived from
the source string. It is not a real concern since the result is guaranteed
to be less than the buffer size due to the MIN operation. However, the
calculation is slightly off anyway. So rewite the logic to kill two birds
with one stone.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>
(cherry picked from commit 502dd95d666c917dc9a031d18d248bbef034dc0f)