fabtests/getinfo: Remove test_caps_regression
The first fi_getinfo returns the caps that are supported by all the
mode bits, including the cap that requires a specific mode bit. The
second fi_getinfo request all caps but only with FI_CONTEXT, so it
could fail with -FI_ENODATA.
This test is checking an incorrect application assumption. Remove it
since it is obsolete.

Signed-off-by: Jessie Yang <jiaxiyan@amazon.com>