prov/opx: Don't ACK packets that were never received
When the very first packet received by an endpoint has a PSN > 0,
and the next expected PSN for that reliability flow is initialized
to 0, the range of PSNs to send ACKs for was incorrectly calculated.
This caused ACKs to be sent for packets that were never actually
received, possibly resulting in hangs or other undesired behavior.

Signed-off-by: Ben Lynam <Ben.Lynam@cornelisnetworks.com>