fabtests: add OOB finalize sync option
ft_finalize_ep is called in all tests and guarantees that no peer exits without
confirming that the other peer is done as well. This is normally done with an
in band send/recv with the FI_TRANSMIT_COMPLETE flag set which guarantees the
send completion won't be generated until the message has completely be sent.
This patch adds an out of band finalize option using the out of band socket if
the user requested out of band syncs.
Typically an in band finalize is sufficient but may be problematic with certain
scenarios. For example, if the provider doesn't support FI_TRANSMIT_COMPLETE (rare)
and for providers using an unreliable protocol underneath (rxd) where the last
underlying acknowledgement may be dropped, which can result in one side exiting
before the other waiting for an acknowldgement retry that will never come

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>