[v2.3.x] fabtests/multi_ep: fix fi_info usage
In the RDM case of the multi_ep test, the test calls fi_getinfo again
for each EP in order to override the source address.
The source address was getting set as the node for the new call but the
FI_SOURCE flag was not set meaning the local source address was getting
resolved into the dst_addr which was never used anyway.
This removes the node parameter with the source address to let the
provider set a source address for us.

This also fixes an issue on cleanup if one of the calls to getinfo fails,
the test would try to cleanup the resources and would segfault because
the fi is NULL.

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>
(cherry picked from commit 1a749cbf0b70a4168edc4f53df20cf21b45d2708)