[v2.3.x] prov/efa: Introduce pke generation counter for protocol path
This commit uses the last 6 bits of the pke pointer for a generation
counter that is incremented every time the packet entry is posted to
rdma-core. When we receive a completion from rdma-core, we verfiy
that the completion is for the latest posted packet.

Signed-off-by: Sai Sunku <sunkusa@amazon.com>
(cherry picked from commit a14460174784e575416290e9e0653c74b218b650)
(cherry picked from commit e83d429493891cad1fee745e4071d06cf8461222)