fabtests/unit: Add av_lookup_good test
Specifically test the av_insert -> av_lookup path.
Use the existing tests as a template and add lookup
capability. This new test will only check that the
"num_good_addr" are correctly inserted into the AV
and when they are looked up with fi_av_lookup, they
are the proper expected values.

There is not a negative version of this test since
provider behavior for fi_av_lookup outside of the
av_count range is undefined. This will be addressed
in a future PR.

Signed-off-by: Zach Dworkin <zachary.dworkin@intel.com>