prov/efa: Do not track rx pkt pool for non-debug build
It is expected to have outstanding rx pkt entries during ep close:

1. for internal rx pkts, we will always keep rx_size number of pkts preposted
2. for user recv rx pkts, it is legimate the user can post more fi_recv without
waiting for completions and then close ep, as there is no way to cancel these recv
today.

For that regard, it is not useful to rely on the dlist iteration to clean the rx pkt pool,
as the memory region should be freed by the bufpool destroy anyway.

Signed-off-by: Shi Jin <sjina@amazon.com>