prov/efa: Decrease the peer robuf size instead of the robuf pool
https://github.com/ofiwg/libfabric/commit/7232f8af12d0a7ad54cc571529e37c006cb3bc92
reduced the size of the robuf pool to 16 while keeping the robuf size
fixed at 8192.

Having different sizes for the the peer pool and the robuf pool does
not make sense becase each peer struct requires a robuf. Instead,
reduce the size of the robuf.

This commit also adds a fabtest with receive window size set to 1 to
check for functional correctness at small receive window sizes

Signed-off-by: Sai Sunku <sunkusa@amazon.com>