prov/efa: Use endpoint->peer hashmap at the AV level
An fi_addr->peer  hashmap at the endpoint level grows
with the number of peers. An endpoint->peer hashmap
grows with the number of endpoints - which would be
better for applications that create many peers but
few endpoints per AV e.g. MPI applications which
only create a single endpoint per AV.

Signed-off-by: Sai Sunku <sunkusa@amazon.com>