prov/lnx: remove force setting DEVICE_ONLY flag
LNX cannot assume the region is device only memory. It has to simply
pass in the user provided flags.
Not using this flag will not allow shm to use IPC copies but the proper
fix is to have the application set the flag (or to use the HMEM hook to
do device information discovery)

Signed-off-by: Zach Dworkin <zachary.dworkin@intel.com>
Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>