[v2.3.x] prov/efa: adjust the memory barrier positions
We need such barrier in two transitions:
1. from wqe write to doorbell ring
2. from doorbell ring to wqe write.

For #2, we currently have the barrier right after the doorbell ring,
which is different from rdma-core implementation that has the doorbell
right in the next wr_start, and according to benchmark such choice
hurts performance on ARM platform. This patch moves the #2 barrier
to the beginning of wr start, before the wqe write, to be consistent
with rdma-core and improves the performance.

Signed-off-by: Shi Jin <sjina@amazon.com>
(cherry picked from commit c4937ed49d476fe72fc97274a2a3dc16b85dfb9a)