[v2.4.x] prov/rxm: prevent segfault on get addr if conn is NULL
When updating the unspec queue address information the peer_context's
connection is sometimes NULL in the case of non FI_DIRECTED_RECV or
non FI_SOURCE. This results in segmentation faults on av_insert so
checking for it before accessing it fixes the issue.

Note: We do not want to fix this by forcing the rx_buf->conn to be set
for all incoming message. We want to avoid an extra lookup/set because
the conn is only needed for FI_SOURCE and FI_DIRECTED_RECV cases.

Signed-off-by: Zach Dworkin <zachary.dworkin@intel.com>
(cherry picked from commit 47e04dcbf9fae46ac1fef3d1edaa72b887c084f6)