prov/util: Update logic for source and dir_recv
The util_srx implementation is overcomplicated for the non-directed
receive case. To simplify this, queuing will be limited to the
unspecified unexpected queue when not using directed receive. This is
because we do not need to enable the per-peer queues unless directed
receive is requested. This makes sure that entries do not get stuck
in the wrong queue because they are all living in the same one. It
also simplifies the lookup for the non-directed receive case.

Signed-off-by: Zach Dworkin <zachary.dworkin@intel.com>