src: Remove iface fallback
If the user is specifically requesting an iface we should always
use that interface.

Do not allow fallback to unrequested iface if the user set
iface is unavailable. This could cause mismatches in runs if
one side selects the user iface and the other does not because
of driver error, network instability, etc.

Signed-off-by: Zach Dworkin <zachary.dworkin@intel.com>